crimecard-backend/requirements.txt
# Core dependencies
tensorflow==2.12.0
keras==2.12.0
transformers==4.30.2

# NLP tools
spacy==3.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz

# Document processing
pdfplumber==0.10.2
python-docx==0.8.11

# Additional dependencies that might be needed
numpy>=1.20.0
pandas>=1.3.0
protobuf==3.20.3